requirements.txt
streamlit
requests==2.31.0
pillow
python-dotenv==1.0.0
